Resolve unit test failures

+++ classes/task/course_restore.php
@@ -148,13 +148,19 @@ class course_restore extends adhoc_task {
             );
             $hookmanager = \core\di::get(hook_manager::class);
             try {
-                $hookmanager->dispatch(new before_restore_precheck($rc, $coursemigration));
+                // The plan is only loaded for valid backups.
+                if ($rc->get_plan() !== null) {
+                    $hookmanager->dispatch(new before_restore_precheck($rc, $coursemigration));
+                }
                 $rc->execute_precheck();
                 $hookmanager->dispatch(new after_restore_precheck($rc, $coursemigration));
                 $rc->execute_plan();
                 $hookmanager->dispatch(new after_restore($rc, $coursemigration));
             } finally {
-                $rc->destroy();
+                // Destroy requires the plan to be set.
+                if ($rc->get_plan() !== null) {
+                    $rc->destroy();
+                }
             }
 
             $coursemigration->set('status', coursemigration::STATUS_COMPLETED)
